+++ libopendmi/include/opendmi/extension/hpe.h
@@ -1,0 +1,16 @@
+//
+// OpenDMI: Cross-platform DMI/SMBIOS framework
+// Copyright (c) 2025-2026, The OpenDMI contributors
+//
+// SPDX-License-Identifier: BSD-3-Clause
+//
+#ifndef OPENDMI_EXTENSION_HPE_H
+#define OPENDMI_EXTENSION_HPE_H
+
+#pragma once
+
+#include <opendmi/extension.h>
+
+extern const dmi_extension_t dmi_hpe_ext;
+
+#endif // !OPENDMI_EXTENSION_HPE_H