+++ libopendmi/src/extension/ibm.c
@@ -1,0 +1,14 @@
+//
+// OpenDMI: Cross-platform DMI/SMBIOS framework
+// Copyright (c) 2025-2026, The OpenDMI contributors
+//
+// SPDX-License-Identifier: BSD-3-Clause
+//
+#include <opendmi/extension/ibm.h>
+
+const dmi_extension_t dmi_ibm_ext =
+{
+    .code     = "ibm",
+    .name     = "IBM",
+    .entities = nullptr
+};