+++ libopendmi/src/extension/lenovo.c
@@ -1,0 +1,14 @@
+//
+// OpenDMI: Cross-platform DMI/SMBIOS framework
+// Copyright (c) 2025-2026, The OpenDMI contributors
+//
+// SPDX-License-Identifier: BSD-3-Clause
+//
+#include <opendmi/extension/lenovo.h>
+
+const dmi_extension_t dmi_lenovo_ext =
+{
+    .code     = "lenovo",
+    .name     = "Lenovo",
+    .entities = nullptr
+};